moving `is_silent` to speech_tools
